requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
nest-asyncio==1.6.0
openai==1.11.1
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
Telethon==1.34.0
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
